Restructured playlist structs
