+++ src/types/mod.rs
@@ -3,6 +3,8 @@ mod delay;
 mod music;
 mod pattern;
 mod playlist;
+mod playlist_item;
+mod prepared_playlist_item;
 mod sequence;
 mod sequence_data;
 mod show;
@@ -10,7 +12,9 @@ mod show;
 pub use self::delay::Delay;
 pub use self::music::Music;
 pub use self::pattern::Pattern;
-pub use self::playlist::{Playlist, PlaylistItem, PreparedPlaylistItem};
+pub use self::playlist::Playlist;
+pub use self::playlist_item::PlaylistItem;
+pub use self::prepared_playlist_item::PreparedPlaylistItem;
 pub use self::sequence::Sequence;
 pub use self::sequence_data::SequenceData;
 pub use self::show::Show;
